Moved staging related code to their own file.
