+++ tests/exporter/test_bundle_update_service.py
@@ -2,8 +2,8 @@ from unittest import TestCase
 
 from mock import Mock, call
 
-from ingest.api.stagingapi import FileDescription
-from ingest.exporter.bundle_update_service import StagingService, Bundle, BundleService, StagingInfo, Exporter
+from ingest.exporter.bundle_update_service import Bundle, BundleService, Exporter
+from ingest.exporter.staging import StagingInfo
 from ingest.exporter.metadata import MetadataResource
 
 
@@ -73,42 +73,6 @@ class BundleTest(TestCase):
         updated_file = bundle.get_file(uuid)
         self.assertEqual(metadata_resource.dcp_version, updated_file.get('version'))
         self.assertEqual('metadata/cell_suspension_x', updated_file.get('content-type'))
-
-
-class StagingServiceTest(TestCase):
-
-    def test_stage_update(self):
-        # given:
-        metadata_resource = MetadataResource(metadata_type='donor_organism',
-                                             uuid='831d4b6e-e8a2-42ce-b7c0-8d6ffcc15370',
-                                             metadata_json={'description': 'test'},
-                                             dcp_version='4.2.1')
-
-        # and:
-        staging_client = Mock(name='staging_client')
-        file_description = FileDescription(['chks0mz'], 'application/json', 'file.name', 1024,
-                                                'http://domain.com/file.url')
-        staging_client.stageFile = Mock(return_value=file_description)
-
-        # and:
-        staging_service = StagingService(staging_client)
-
-        # when:
-        staging_area_uuid = '7455716e-9639-41d9-bff9-d763f9ee028d'
-        staging_info = staging_service.stage_update(staging_area_uuid, metadata_resource)
-
-        # then:
-        self.assertTrue(type(staging_info) is StagingInfo,
-                        'stage_update should return StagingRecord.')
-        self.assertEqual(metadata_resource.uuid, staging_info.metadata_uuid)
-        self.assertEqual(file_description.name, staging_info.file_name)
-        self.assertEqual(file_description.url, staging_info.cloud_url)
-
-        # and:
-        staging_client.stageFile.assert_called_once_with(staging_area_uuid,
-                                                         metadata_resource.get_staging_file_name(),
-                                                         metadata_resource.metadata_json,
-                                                         metadata_resource.metadata_type)
 
 
 class BundleServiceTest(TestCase):
